Move RichPresence to core package
