+++ bundles/fr.kazejiyu.discord.rpc.integration/src/fr/kazejiyu/discord/rpc/integration/extensions/impl/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Default implementations for {@link fr.kazejiyu.discord.rpc.integration.extensions.RichPresence RichPresence}
+ * Default implementations for {@link fr.kazejiyu.discord.rpc.integration.core.RichPresence RichPresence}
  * interface.
  * 
  * @author Emmanuel CHEBBI
